refactor: Changes structure, minor edits in business-logic, adds docs
