+++ main-service/src/main/java/ru/acceleration/store/dto/product/NewProductDto.java
@@ -1,7 +1,6 @@
-package ru.acceleration.store.dto;
+package ru.acceleration.store.dto.product;
 
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,15 +9,13 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class NewSaleDto {
+public class NewProductDto {
+
+    @NotBlank
+    @Size(max = 30)
+    private String vendorCode;
 
     @NotBlank
     @Size(max = 100)
     private String name;
-
-    @NotBlank
-    private String quantity;
-
-    @Positive
-    private Long promotionId;
 }
