+++ main-service/src/main/java/ru/acceleration/store/dto/product/ProductFullDto.java
@@ -1,4 +1,4 @@
-package ru.acceleration.store.dto;
+package ru.acceleration.store.dto.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
