+++ main-service/src/main/java/ru/acceleration/store/dto/sale/NewSaleDto.java
@@ -1,6 +1,7 @@
-package ru.acceleration.store.dto;
+package ru.acceleration.store.dto.sale;
 
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,13 +10,15 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class NewProductDto {
-
-    @NotBlank
-    @Size(max = 30)
-    private String vendorCode;
+public class NewSaleDto {
 
     @NotBlank
     @Size(max = 100)
     private String name;
+
+    @NotBlank
+    private String quantity;
+
+    @Positive
+    private Long promotionId;
 }
