+++ main-service/src/main/java/ru/acceleration/store/dto/sale/SaleDto.java
@@ -1,4 +1,4 @@
-package ru.acceleration.store.dto;
+package ru.acceleration.store.dto.sale;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
