+++ main-service/src/main/java/ru/acceleration/store/dto/user/UserDto.java
@@ -1,4 +1,4 @@
-package ru.acceleration.store.dto;
+package ru.acceleration.store.dto.user;
 
 import lombok.*;
 import lombok.experimental.FieldDefaults;
