+++ main-service/src/main/java/ru/acceleration/store/mapper/ProductMapper.java
@@ -2,9 +2,9 @@ package ru.acceleration.store.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
-import ru.acceleration.store.dto.NewProductDto;
-import ru.acceleration.store.dto.ProductFullDto;
-import ru.acceleration.store.dto.ProductShortDto;
+import ru.acceleration.store.dto.product.NewProductDto;
+import ru.acceleration.store.dto.product.ProductFullDto;
+import ru.acceleration.store.dto.product.ProductShortDto;
 import ru.acceleration.store.model.Product;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
