+++ main-service/src/main/java/ru/acceleration/store/mapper/PromotionMapper.java
@@ -2,7 +2,7 @@ package ru.acceleration.store.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
-import ru.acceleration.store.dto.PromotionDto;
+import ru.acceleration.store.dto.promotion.PromotionDto;
 import ru.acceleration.store.model.Promotion;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
