+++ main-service/src/main/java/ru/acceleration/store/mapper/SaleMapper.java
@@ -3,8 +3,8 @@ package ru.acceleration.store.mapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
-import ru.acceleration.store.dto.NewSaleDto;
-import ru.acceleration.store.dto.SaleDto;
+import ru.acceleration.store.dto.sale.NewSaleDto;
+import ru.acceleration.store.dto.sale.SaleDto;
 import ru.acceleration.store.model.Sale;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
