+++ main-service/src/main/java/ru/acceleration/store/mapper/UserMapper.java
@@ -2,8 +2,8 @@ package ru.acceleration.store.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
-import ru.acceleration.store.dto.UserCreateDto;
-import ru.acceleration.store.dto.UserDto;
+import ru.acceleration.store.dto.user.UserCreateDto;
+import ru.acceleration.store.dto.user.UserDto;
 import ru.acceleration.store.model.User;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
