+++ main-service/src/main/java/ru/acceleration/store/service/product/ProductServiceImpl.java
@@ -1,11 +1,11 @@
-package ru.acceleration.store.service;
+package ru.acceleration.store.service.product;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.acceleration.store.dto.NewProductDto;
-import ru.acceleration.store.dto.ProductFullDto;
-import ru.acceleration.store.dto.ProductShortDto;
+import ru.acceleration.store.dto.product.NewProductDto;
+import ru.acceleration.store.dto.product.ProductFullDto;
+import ru.acceleration.store.dto.product.ProductShortDto;
 import ru.acceleration.store.mapper.ProductMapper;
 import ru.acceleration.store.model.Product;
 import ru.acceleration.store.repository.ProductRepository;
@@ -29,17 +29,23 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
+    public ProductFullDto getProductById(Long productId) {
+        Product foundProduct = productRepository.getExistingProduct(productId);
+        return productMapper.toProductFullDto(foundProduct);
+    }
+
+    /**
+     * Метод отдает все товары, которые входят в баннер
+     *
+     * @param promotionId Id баннера
+     * @return список товаров, входящих в подборку баннера/акции
+     */
+    @Override
     public List<ProductShortDto> productsInPromotion(Long promotionId) {
         return productRepository.findProductByPromotion(promotionId)
                 .stream()
                 .map(productMapper::toProductShortDto)
                 .collect(Collectors.toList());
-    }
-
-    @Override
-    public ProductFullDto getProductById(Long productId) {
-        Product foundProduct = productRepository.getExistingProduct(productId);
-        return productMapper.toProductFullDto(foundProduct);
     }
 
     @Override
