+++ main-service/src/main/java/ru/acceleration/store/service/promotion/PromotionServiceImpl.java
@@ -1,16 +1,17 @@
-package ru.acceleration.store.service;
+package ru.acceleration.store.service.promotion;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.acceleration.store.dto.ProductShortDto;
-import ru.acceleration.store.dto.PromotionDto;
+import ru.acceleration.store.dto.product.ProductShortDto;
+import ru.acceleration.store.dto.promotion.PromotionDto;
 import ru.acceleration.store.mapper.PromotionMapper;
 import ru.acceleration.store.model.Promotion;
 import ru.acceleration.store.model.Sale;
 import ru.acceleration.store.repository.PromotionRepository;
 import ru.acceleration.store.repository.SaleRepository;
+import ru.acceleration.store.service.product.ProductService;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -32,6 +33,12 @@ public class PromotionServiceImpl implements PromotionService {
         return promotionMapper.toPromotionDto(newPromotion);
     }
 
+    /**
+     * Метод удаляет баннер.
+     * Если к баннеру привязаны акции (Sale), то они не удаляются, но отвязываются от баннера
+     *
+     * @param promotionId Id баннера/промо-акции
+     */
     @Transactional
     @Override
     public void deletePromotion(Long promotionId) {
@@ -64,6 +71,13 @@ public class PromotionServiceImpl implements PromotionService {
         return productService.productsInPromotion(promotionId);
     }
 
+    /**
+     * Метод меняет название у баннера.
+     *
+     * @param promotionId Id баннера
+     * @param newPromotionDto Id данные для изменения
+     * @return измененный банер
+     */
     @Override
     public PromotionDto editPromotion(Long promotionId, PromotionDto newPromotionDto) {
         Promotion existingPromotion = promotionRepository.getExistingPromotion(promotionId);
@@ -74,6 +88,6 @@ public class PromotionServiceImpl implements PromotionService {
 
     @Override
     public Promotion getPromotionById(Long promotionId) {
-        return promotionRepository.getReferenceById(promotionId);
+        return promotionRepository.getExistingPromotion(promotionId);
     }
 }
