+++ main-service/src/main/java/ru/acceleration/store/service/user/UserService.java
@@ -1,0 +1,10 @@
+package ru.acceleration.store.service.user;
+import org.springframework.stereotype.Service;
+import ru.acceleration.store.dto.user.UserCreateDto;
+import ru.acceleration.store.dto.user.UserDto;
+
+@Service
+public interface UserService {
+
+    UserDto postUser(UserCreateDto userCreateDto);
+}
