+++ main-service/src/test/java/ru/acceleration/store/SaleControllerTests.java
@@ -10,9 +10,9 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-import ru.acceleration.store.dto.NewProductDto;
-import ru.acceleration.store.dto.NewSaleDto;
-import ru.acceleration.store.dto.PromotionDto;
+import ru.acceleration.store.dto.product.NewProductDto;
+import ru.acceleration.store.dto.sale.NewSaleDto;
+import ru.acceleration.store.dto.promotion.PromotionDto;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -29,6 +29,7 @@ public class SaleControllerTests {
 
     NewSaleDto saleCreateDto = new NewSaleDto("Такая вот скидка", "200", null);
     NewSaleDto saleCreateDtoWithPromotion = new NewSaleDto("Такая вот скидка в подборке", "200", 1L);
+    NewSaleDto saleCreateDtoWithWrongPromotion = new NewSaleDto("Такая вот скидка в несуществующей подборке", "200", 10L);
     NewProductDto productCreateDto = new NewProductDto("XY73GS33", "Apple iPhone 13 Pro Max 256GB");
     PromotionDto promotionCreateDto = new PromotionDto(null, "Компьютеры со скидкой до -30%", null);
 
@@ -43,23 +44,23 @@ public class SaleControllerTests {
     }
 
     @Test
-    void postProductTestWithoutPromotion() throws Exception {
+    void postSaleWithoutPromotion() throws Exception {
         mockMvc.perform(MockMvcRequestBuilders
                         .post("/products")
                         .content(mapper.writeValueAsString(productCreateDto))
                         .contentType(MediaType.APPLICATION_JSON));
         mockMvc.perform(MockMvcRequestBuilders
-                        .post("/product/1/sale")
+                        .post("/product/2/sale")
                         .content(mapper.writeValueAsString(saleCreateDto))
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.name").value("Такая вот скидка"))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.quantity").value("200"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.productId").value(1))
+                .andExpect(MockMvcResultMatchers.jsonPath("$.productId").value(2))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.promotionId").doesNotExist());
     }
 
     @Test
-    void postProductTestWithPromotion() throws Exception {
+    void postSaleWithPromotion() throws Exception {
         mockMvc.perform(MockMvcRequestBuilders
                 .post("/products")
                 .content(mapper.writeValueAsString(productCreateDto))
@@ -76,5 +77,18 @@ public class SaleControllerTests {
                 .andExpect(MockMvcResultMatchers.jsonPath("$.quantity").value("200"))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.productId").value(1))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.promotionId").value(1));
+    }
+
+    @Test
+    void postSaleWithPromotionWhichDoesntExist() throws Exception {
+        mockMvc.perform(MockMvcRequestBuilders
+                .post("/products")
+                .content(mapper.writeValueAsString(productCreateDto))
+                .contentType(MediaType.APPLICATION_JSON));
+        mockMvc.perform(MockMvcRequestBuilders
+                        .post("/product/1/sale")
+                        .content(mapper.writeValueAsString(saleCreateDtoWithWrongPromotion))
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isNotFound());
     }
 }
